HIPCMS-431 added permissions

+++ src/Api/Controllers/AnnotationController.cs
@@ -1,21 +1,23 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 using Api.Data;
 using Microsoft.Extensions.Logging;
 using Api.Managers;
 using Microsoft.AspNetCore.Mvc;
 using Api.Models.AnnotationTag;
+using Api.Permission;
+using Api.Utility;
 
 namespace Api.Controllers
 {
     public class AnnotationController : ApiController
     {
-        AnnotationTagManager tagManager;
+        private AnnotationTagManager tagManager;
+        private AnnotationPermissions annotationPermissions;
+
         public AnnotationController(CmsDbContext dbContext, ILoggerFactory loggerFactory) : base(dbContext, loggerFactory)
         {
             tagManager = new AnnotationTagManager(dbContext);
+            annotationPermissions = new AnnotationPermissions(dbContext);
         }
 
         #region GET
@@ -59,9 +61,13 @@ namespace Api.Controllers
         [HttpPost("Tags")]
         [ProducesResponseType(typeof(IActionResult), 200)]
         [ProducesResponseType(typeof(void), 500)]
+        [ProducesResponseType(typeof(void), 403)]
         [ProducesResponseType(typeof(void), 400)]
         public IActionResult Post(AnnotationTagFormModel tag)
         {
+            if (!annotationPermissions.IsAllowedToCreateTags(User.Identity.GetUserId()))
+                return Forbid();
+
             if (ModelState.IsValid)
             {
                 var result = tagManager.AddTag(tag);
@@ -75,9 +81,13 @@ namespace Api.Controllers
         // Post api/Annotation/Tags/:parentId/ChildTags/:childId
         [HttpPost("Tags/{parentId}/ChildTags/{childId}")]
         [ProducesResponseType(typeof(void), 200)]
+        [ProducesResponseType(typeof(void), 403)]
         [ProducesResponseType(typeof(void), 400)]
         public IActionResult Post(int parentId, int childId)
         {
+            if (!annotationPermissions.IsAllowedToEditTags(User.Identity.GetUserId()))
+                return Forbid();
+
             bool success = tagManager.AddChildTag(parentId, childId);
             if (success)
                 return Ok();
@@ -91,9 +101,13 @@ namespace Api.Controllers
         // Put api/Annotation/Tags/:Id
         [HttpPut("Tags/{Id}")]
         [ProducesResponseType(typeof(void), 200)]
+        [ProducesResponseType(typeof(void), 403)]
         [ProducesResponseType(typeof(void), 404)]
         public IActionResult PutTag(int Id, AnnotationTagFormModel model)
         {
+            if (!annotationPermissions.IsAllowedToEditTags(User.Identity.GetUserId()))
+                return Forbid();
+
             var success = tagManager.EditTag(model, Id);
             if (success)
                 return Ok();
@@ -107,8 +121,12 @@ namespace Api.Controllers
         // Delete api/Annotation/Tags/:Id
         [HttpDelete("Tags/{Id}")]
         [ProducesResponseType(typeof(void), 200)]
+        [ProducesResponseType(typeof(void), 403)]
         public IActionResult Delete(int Id)
         {
+            if (!annotationPermissions.IsAllowedToCreateTags(User.Identity.GetUserId()))
+                return Forbid();
+
             bool success = tagManager.DeleteTag(Id);
             if (success)
                 return Ok();
@@ -117,8 +135,12 @@ namespace Api.Controllers
 
         [HttpDelete("Tags/{parentId}/ChildTags/{childId}")]
         [ProducesResponseType(typeof(void), 200)]
+        [ProducesResponseType(typeof(void), 403)]
         public IActionResult DeleteChildOf(int parentId, int childId)
         {
+            if (!annotationPermissions.IsAllowedToEditTags(User.Identity.GetUserId()))
+                return Forbid();
+
             bool success = tagManager.RemoveChildTag(parentId, childId);
             if (success)
                 return Ok();
